move all problem page
