--- src/app/problems/all/[page]/types.ts
@@ -1,1 +1,0 @@
-export type { ProblemList } from "@utils/api/getAllProblem";
