+++ src/features/problems/components/ProblemList/ProblemScore.tsx
@@ -2,7 +2,8 @@ import { cookies } from "next/headers";
 
 import { Score } from "@/components";
 import { getMemoUser } from "@/features/auth";
-import { getMemoProblemsScore } from "@/features/problems";
+
+import { getMemoProblemsScore } from "../../queries";
 
 export const ProblemScoreText = ({ text }: { text: string }) => (
   <p className="font-semibold text-slate-600 dark:text-slate-200">{text}</p>
