+++ src/features/problems/components/index.ts
@@ -1,3 +1,5 @@
+export * from "./CreateProblem";
 export * from "./Editor";
 export * from "./InfoTable";
+export * from "./ProblemList";
 export * from "./RecalcScore";
