+++ src/features/problems/queries/index.ts
@@ -1,4 +1,6 @@
+export * from "./getAllProblem";
 export * from "./getProblem";
+export * from "./getProblemCount";
 export * from "./getProblemScore";
 export * from "./getProblemsScore";
 export * from "./recalcProblemScore";
