+++ src/features/problems/types/index.ts
@@ -1,5 +1,16 @@
 import type { Judge } from "@utils/shared/parseJudge";
 
+import type { Languages } from "@/types";
+
+export interface ProblemInfo {
+  dpid: number;
+  code: string;
+  name: string;
+  language: Languages;
+}
+
+export type ProblemList = ProblemInfo[];
+
 export interface DetailedProblemInfo {
   code: string;
   name: string;
