--- src/shared/utils/api/index.ts
@@ -1,4 +1,0 @@
-import "server-only";
-export { getAllProblem } from "./getAllProblem";
-export { getProblemCount } from "./getProblemCount";
-export { getSubmissionInfo, type SubmissionInfo } from "./getSubmissionInfo";
